refactor: Seperated cards into seperate files
